New Roundtrip Integration (#8)
* VIS directory setup as discussed in https://docs.google.com/document/d/1NU5VOPzlyrmRNZMmnSHfMuZ0sD71vVKdV4sHJ44QtxM/edit\#

* Added roundtrip files to vis/external subdirectory; removed empty example files which can be re-added by with content. Modified cct to be built with webpack and loaded using new roundtrip.

* Update to hatchet.vis.external for roundtrip retrieval

* Added patch to fix quote closure problem in data transfer

* Reinstated call back registration for roundtrip watching

* Removed empty files, added build and watch commands to package.json, webpack minifies code by default, added readme

* moved old rountrip to roundtrip_old for merging

* Rebase complete

* fix formatting issues

* Moved hatcehet to top level external directory.

* Added import catcher for when hatchet is run without jupyter.
