+++ hatchet/external/roundtrip/NOTICE
@@ -1,0 +1,2 @@
+This work was supported by the United States Department of Defense through DTIC Contract FA8075-14-D-0002-0007, 
+the National Science Foundation under NSF III-1656958, and the Department of Energy under DE-SC0022044.
