+++ hatchet/external/roundtrip/requirements.txt
@@ -1,0 +1,2 @@
+beautifulsoup4
+pycountry_convert
