+++ hatchet/external/roundtrip/roundtrip/__init__.py
@@ -1,0 +1,4 @@
+# Copyright 2021-2022 HDC Lab and Roundtrip Project Developers.
+# See the top-level LICENSE file for details.
+#
+# SPDX-License-Identifier: MIT
