+++ hatchet/external/roundtrip/roundtrip/version.py
@@ -1,0 +1,2 @@
+__version_info__ = ("0", "2", "1")
+__version__ = ".".join(__version_info__)
