+++ hatchet/vis/external/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from .roundtrip.roundtrip.manager import Roundtrip
